Further simplify StrictCountCollector, and fix a bug
Summary:
The StrictCountCollector logic seems like it was a bit weird - we were independently
tracking strict vs unsafe comment lines, and not only that but also we were
treating an ignore-all-errors directive as unsafe rather than ignore-all which is
just incorrect.

We can do much better by just trackign the mode and any explicit comment line. This
will behave very slightly differently if there are duplicates, but since
duplicates won't type-check it's silly to be worrying much about the semantics
of our strictness-collection in the presence of duplicates.

We were also tracking counts, as if this visitor were going to be run across many
modules. But that never happens, so that was all dead code - I've removed the counts
entirely and no use case is affecteced.

Reviewed By: grievejia

Differential Revision: D44399006

fbshipit-source-id: 2489b6e9229fbb063c494f2d588d5bfd7b2192ac