Extract context-tracking into its own class
Summary:
One thing I don't love about libcst visitors is how we tend to want to smash a bunch
of utilities and mutable state all into one place.

I don't necessarily have a problem with this complexity *existing*, because after all
the visitor has to do everything in the end. But having everything live at the same
layer of the codebase makes it hard to see which methods are related to one another,
and how state mutation relates to various actual behaviors.

Extracting the state into a context class that is mutable and has explicit
update + query functions does not remove any complexity, but it lets us make the
coupling between updates and queries explicit, giving semantic names for each
operation that the higher level can refer to that it's clear *why* this state tracking
is needed, which isn't very clear when we do it all in one layer.

Reviewed By: grievejia

Differential Revision: D44377636

fbshipit-source-id: c542d2a06236e3c7de5ea13c621dc70af701952a