Thin down the fixme counting tests a bit
Summary:
There were more test cases than needed in my opinion, which is
annoying if I want to re-implement the suppression finding logic (which
I do, I'll explain why in a later diff) while reusing the tests.

So simplify a bit. Aggregations are isolated into the first two
test cases, everything after that is verifying that the regex handling
works properly with the comment in various places.

Reviewed By: grievejia

Differential Revision: D44398565

fbshipit-source-id: 90f4057e91968779a7abc6ee97323242041a24d6