Adds support for checking python version in models (#720)
Summary:
Adds support for checking python versions in pysa models by extending the model parser to parse if-conditions as well.

Previously all if-conditions in the modes file were discarded with a model verification error of type Unsupportedstatement. Extend the AST logic to check for the presence of if conditions and honour them. Currently, we can check against the python version in the system provided to the taintAnalysis Command and while running the server by accessing the configuration.

This allows us to write if conditions along with else if blocks if needed in model files of the form:
`if sys.version <operator> (<python_major_version>, <python_minor_version>, <python_micro_version>):`
The models in the if body or the else body will be parsed only if the condition evaluates to true. sys.version is the only allowed left hand expression, all other expressions lead to a model verification error.

Nesting of if conditions are also supported. All comparison operators except in and not in are supported. If the right side expression is not a tuple, the if condition is ignored and a model verification error is emitted.

Update existing tests and ModelParser.parse invocations to pass in the python_version. Adds new unit tests as well.

Also adds documentation for the same.

Pull Request resolved: https://github.com/facebook/pyre-check/pull/720

Test Plan:
- Build the project from source using dune
- cd a repository down
- For ease, modify the model in documentation/pysa_tutorials/exercise1 to add an if condition with a comparison operator to it, eg:
```
if sys.version < (2,1,0):
    django.http.request.HttpRequest.GET: TaintSource[CustomUserControlled] = ...
    def eval(__source: TaintSink[CodeExecution], __globals, __locals): ...
```
- Create the following .pyre_configuration file:
```
{
    "source_directories": ["pyre-check/documentation/pysa_tutorial/exercise1/"],
    "search_path": ["./stubs/stubs"],
    "taint_models_path": "./pyre-check/documentation/pysa_tutorial/exercise1/"
}
```
- run pysa and see that the model is not getting detected. (No issues):

<img width="1440" alt="Screenshot 2023-03-20 at 11 44 18 AM" src="https://user-images.githubusercontent.com/8947010/226263061-10e4d90c-819c-498b-a052-cad26e4627bf.png">

- change the model such that the if condition evaluates to true
```
if sys.version > (2,1,0):
    django.http.request.HttpRequest.GET: TaintSource[CustomUserControlled] = ...
    def eval(__source: TaintSink[CodeExecution], __globals, __locals): ...
```
- run pysa again and see that the model does get detected (1 issue found):

<img width="1440" alt="Screenshot 2023-03-20 at 11 45 13 AM" src="https://user-images.githubusercontent.com/8947010/226263166-31667fec-d07f-423a-adb8-de91d5def179.png">

- change the if condition to something else, for example, `if (2,1,0):` to see a model verification error:

<img width="1440" alt="Screenshot 2023-03-20 at 11 46 49 AM" src="https://user-images.githubusercontent.com/8947010/226263290-d6f8747e-fb5b-4113-a466-f154b629bbba.png">

- run pysa on delibrately_vulnerable_app to make sure this doesn't break anything (github action)
- unit tests and github actions

Fixes: https://github.com/MLH-Fellowship/pyre-check/issues/83

Reviewed By: tianhan0

Differential Revision: D44422810

Pulled By: arthaud

fbshipit-source-id: 40c7d411af66112da19dead4ee2a43e6b65c0d54