Include `opam init` call in setup script
Summary:
The rust dependency we are adding requires opam init be set up with `--disable-sandboxing` and our current setup flow (setup.sh) requires running `opam init` beforehand. The documentation doesn't say we need to run `opam init` with the arguments provided here. To make it easier we should just rely on setup.sh to call init so these args are not forgotten.

~~Potentially contreversial: This diff also adds `--shell-support` to `Automatically setup the user shell configuration for opam, e.g. adding a line to the `~/.profile' file.` This means we won't have to run eval around all opam commands.~~

I'll update the documentation to include the incorrect usage after this + next diff lands.

Reviewed By: grievejia

Differential Revision: D44488934

fbshipit-source-id: 5bdbda5afa2c0e8699ef378d7cfdff4c11f4d62f