Better molding of taint trees
Summary:
# Context
When building a model, we apply a tree molding operation that eleminates some duplicate paths.
This is done by building a mold, which is a taint tree similar to the existing one, but only containing essential information (removing breadcrumbs, for instance). Since the abstract tree domain automatically prunes taint from branches if is already exists in the root, this will create a pruned tree.
Finally, we use the pruned tree to shape the original taint tree, collapsing any branches that are not present in the mold.
This is beneficial for performance, while costing a little bit of precision.
In general, this seems to remove duplicate paths in traces (e.g, paths `foo -> bar formal(arg)` and `foo -> bar formal(arg)[field]` are merged if they have the same taint kind)

# Problem
The way we build our mold is buggy and error-prone. We iterate on the whole taint tree and remove bits on information we do not want. It is very easy to forget to remove things when implementing a new feature. As an example, we forgot to remove `MultiSourceIssueHandle` from it.

# Solution
Let's rewrite completely `essential`, the function that builds that mold.
We now simply iterate on things we want to keep and create a fresh taint tree from it.

Reviewed By: alexkassil

Differential Revision: D44509768

fbshipit-source-id: 873cc383a4d130bf36a3d02b9620a412093c2030