Error on assignment to global name with attribute
Summary:
**Background**: When erroring on assignments, we currently only check to see if the base named identifier we're assigning to is a global.

Globals, however, can be imported from other modules or exist as classes (if you consider class attributes). These can exist as named attribute expressions, as well as the named identifier we were checking before, so this diff adds a check to ensure we're erroring if the name + attribute is found to be a global.

Reviewed By: pradeep90

Differential Revision: D44473459

fbshipit-source-id: 4efcc3108895afd53cbeb347878dfa3b2e1fb132