dataclasses_json_ify coverage_data code
Summary:
Convert all the classes in `coverage_data.py` to use snake-case
format dataclasses json. This is by far the lowest-touch way to build
output data.

The new `Location` class is used because dataclasses-json, much like the
yojson ppx in ocaml, is "viral" in that by far the easiest way to us it
is to use it "all the way down the stack". We don't control the
`libcst.CodeRange` class so it's nicer to use our own representation.

A side benefit of this is that the new location representation is a bit more
flat, which will make it easier to translate into tabular data for
hive (which is where all this data needs to go if we want it to power
doashbaords).

Reviewed By: grievejia

Differential Revision: D44587340

fbshipit-source-id: 4df9e930185ab99999b9c7cbcc8678e162947b59