Switch coverage_data to collect in static functions
Summary:
I want to add a collector function for `AnnotationCollector`, but
I realized the pattern of using a `collect` method on the classes does
not really allow this. The reason is that we're using inheritance to
inherit behavior and attributes in the `AnnotationCountCollector`
used by `statistics.py`, but overriding `collect` with a different
output type isn't valid in terms of static typing.

(The collector I was trying to add when I realized this problem is
provided by the `collect_function_annotations` function defined
in this commit - that code cannot be implemented as a `collect`
method without introducing type errors).

This suggests that, given our desire to inherit behavior but not
the output data interface, the collect methods don't really belong
on the classes at all; instead I'm making them static module-level
functions.

I think the callsites are easier to read than intantiating a class
and calling collect anyway; the fact that there's a visitor involved
is really an implementation detail rather than something downstream
code really needs to be exposed to.

I'm not updating the `statistics.py` code, which is in maintenance
mode anyway so having collect as a method should be fine there for
the short remaining lifetime of the code.

Reviewed By: grievejia

Differential Revision: D44587343

fbshipit-source-id: efb454c1e39abb2a8f43b22d2fef75be63f4bd7f