Avoid erroring after non-mutable call on global or attribute
Summary:
**Background**: For global leak detection, we currently have the ability to detect writes to globals and their class attributes -- e.g. `MyClass.x = 5`, `MyClass.x.append(123)`, `MyClass.x.y[0].append(1)`).

We would now like to avoid returning false positives when doing mutations on classes that were instantiated in the same line (e.g. `MyClass().x.append(123)`) or on mutations we do from values that are returned from non-mutable calls to globals.

We also, however, want to make sure we still error on things we know can still produce a global leak:
1. Cases where we return a class from a function (e.g. `def foo() -> Type[MyClass]: ...`) can be a sneaky way to mutate class attributes.
2. Cases where we are fairly confident a function call is going to return a member of a global, where mutating that member counts as mutating the global (e.g. `my_list[5] = 1`).

**Context**: We're currently able to support writes to instance attributes (e.g. `MyClass().x = 5`), but the previous diff adds functionality to detect globals that have a qualifier.

The upcoming diff will focus on fixing behavior around globals in mutable expression call arguments.

In the next few weeks, I'd like to redo some of how this works to get rid of some weaknesses this approach takes.

Reviewed By: pradeep90

Differential Revision: D44468584

fbshipit-source-id: 457ff08f697d0c5d26db3bc3bdbf92a2d08adb29