Set up service restarting.
Summary:
For reasons that aren't clear to me (it can be tricky
to set up great logging in a 3-part distributed system where
a flask worker shells out to the client which connects to a
binary with 10 separate workers), our playground is crashing
about once a month; somethign goes wrong in one worker and then
gunicorn shuts all the workers down.

Realistically this isn't a big problem given how rare it is,
and I don't have time to investigate in depth. Since our playground
flask server is just a systemd daemon, we can just set a restart
directive.

Reviewed By: grievejia

Differential Revision: D44591760

fbshipit-source-id: 43c8df14cb9528009eb749bd7f9a4fff2ad7c1d6