Update testing framework to error on type errors
Summary:
**Background**: In a relatively recent diff, it was discovered that some tests were failing silently because of type or format errors, and not actually producing errors when the global leak check was failing.

To make it more difficult for this to occur again in the future, type checking functionality has been added to the testing framework, and the test will fail because of an error (instead of a regular test failure, which will not show the location of the failing test) when there are issues with the test source.

This diff also adds the ability to specify additional modules, so that importing functions and globals from other modules can be tested properly.

Reviewed By: pradeep90

Differential Revision: D44520770

fbshipit-source-id: f8f2d54bb9fc217685262426912b0e67154eb3b8