Add a note that `statistics.py` is in maintenance mode
Summary:
The `pyre report` command is now functional, and produces data
that replaces everything we care about in `pyre statistics` (by
"we care about" I just mean it drops globals and attributes, which
we've decided are not worth counting given that they are prone to
both FPs and FNs).

We still cannot remove `statistics.py` for a while yet but we
should note that it is in maintenance mode in the docstring
of both the module (which is developer-facing) and the click
entrypoint (which becomes the help description).

Reviewed By: grievejia

Differential Revision: D44591239

fbshipit-source-id: 8c51645df9d0bcfe13b1470b575a54080dfe1776