Wire up a (placeholder) `pyre report` command
Summary:
In order to power upgraded coverage data infrastructure, I want
to implement a new command that will handle producing all kinds of
coverage-related information (suppressions, interface coverage,
expression coverage).

Why a new command?

Under the hood the new command will largely reuse the same logic
as `pyre statistics` and eventually expression-level coverage,
but we want the ability to rebuild the interface without backward

(which I've started refactoring) that we use for gathering
statistics and coverage today, but I want the freedom to iterate
rapidly on the interface, so reusing the existing `pyre statistics`
command is not an option - that has Sandcastle code coupled to its
interface, and therefore isn't safe to rapidly rewrite.

Reviewed By: kinto0

Differential Revision: D44587341

fbshipit-source-id: 45bb51f3e10b3cbbd2fde7e3cb4033c696889284