Emit error when global (or global attribute) is in return position
Summary:
**Background**: We would like to prevent globals (or their attributes/members) from being returned from functions or method calls, since it's an easy way to miss leaks when running an intraprocedural analysis. We still want to allow classes to be returned (in the case of the factory pattern).

Currently, no extra analysis is performed on the `reachable_globals` in return statement position, so we can achieve the desired results by erroring on *some* of those reachable globals.

We decide to error on a reachable global if it's not a class. To determine this once we reach the top-level statement, we need to include information on the global's type when creating the reachable global, so a new struct has been created for that.

Reviewed By: pradeep90

Differential Revision: D44669945

fbshipit-source-id: c71b822aafef1e83e0cbf9ae8be92a0acc7f8b64