Add multiprocessing to `pyre report`
Summary:
On trunk, `pyre statistics` actually is about 50% *slower* than a full check on
Pyre itself, even including Buck time, because LibCST visiting is so slow. The
new `pyre report` is a little faster becasue we do all the suppression visiting
in one pass, which helps, but it's still slower than a full check.

A lot of this is that doing work in Python and with LibCST is inherently slow
compared to the backend, but another huge factor is that the backend parallelizes
work. Let's use multiprocessing to do the same in the frontend.

Reviewed By: pradeep90

Differential Revision: D44607397

fbshipit-source-id: 163c2997b22c39f66a345c7714e35e795c06665f