Rename annotation_kind to annotation_status
Summary:
The name "annotation kind" suggests that this bit of data is some how a
kind of annotation, which is not the case - it's a classification of the
function depending on how annotated it is.

Using a better name will make it easier to have well-named columns
in downstream tables, so it's worth trying to get this right.

Reviewed By: pradeep90

Differential Revision: D44639545

fbshipit-source-id: 24b0dd9ae3add0df4cee0309387bf037b2f3f195