add _create_configuration in external pyre client
Summary:
This is a fairly simple diff. We will be replacing the commands/ APIs with internal_configuration instead of configuration_module. By making these config creations use a helper function, we can just change that function instead of every caller. The diff changing the commands/ APIs' configuration will be very large so this diff is nice to have.

This diff also makes another _create_and_check_configuration to _check_open_source_version since the caller won't have access to the `configuration_module` configuration after calling `_create_configuration`. This is because future diffs are migrating to the `frontend_configuration` configuration.

Reviewed By: grievejia

Differential Revision: D44370403

fbshipit-source-id: 22e061caebb772fe6a3833bab07af19c64b25059