Abbreviate names of data in the report
Summary:
Something I realized as I started looking into naming the
tables that will hold pyre report data is that the names of things
in the report are kind of long - why "mode info" instead of just "mode"?
Why "function_annotations" instead of just "funcitons"? In the context
of a "pyre report", the longer names don't carry much value.

This wasn't as obvious to me when I was writing the code because low-level
concerns, such as "well a Mode is the enum, the data about the mode plus
the line number has to have a different name like ModeInfo" tend to dominate.

But once I'm looking at consuming the actual json dump it becomes clear
that some of the names are too verbose. Since I don't want to have hive tables
or hive columns with longer names than make sense, I wanted to shorten the names
now before I start writing the upload scripts.

Differential Revision: D44717124

fbshipit-source-id: a8bbd9db659f15de3a8a16d53564569bfe81753c