Add all globals to the call graph as a global
Summary:
To be able to catch a canonical flow through globals D43946168, we need to have globals present in the call graph.

A performance test here D44387242 shows no performance regression when we don't filter for globals that have a pre initialized model.

I believe the reason that we have all these builtings (int, print, isinstance, etc), added as a "global" in these unit tests is because we aren't passing in builtin models to our test call graph construction. When pysa is ran fully, these aren't treated as globals, but as actual models. This part is RFC because I am unsure how to verify this.

Reviewed By: arthaud

Differential Revision: D44542352

fbshipit-source-id: 8ced9995efab6fc0a2093452bf9382d0805785d4