Extract attach trace to top-level
Summary:
Context: We want to separate out callgraph utilities to allow other tools (for example to collect readonly adoption metrics) to use the same setup to walk through transitive callees given an entrypoint. The following set of diffs accomplish this separation.

This diff extracts `attach_trace_to_query_results` to top level since it is not using any attributes of callgraph and updates the related tests.

Reviewed By: connernilsen

Differential Revision: D44712233

fbshipit-source-id: e2b84ce94ecc9227f69f7fae669b2365b2650bdd