Refactor to extract functionality relating to pyre queries
Summary:
Context: Currently `analyze_script` has the functionality to take a list of entrypoints and carry out leak analysis on all the transitive callees from the entrypoints. We would like an additional option to carry out leak analysis on any list of functions. As an example, this addition would help us analyze all functions in IG to gather upper bound on number of leaks.

The following set of diffs will implement this functionality. In this diff, I extract out functionality relating to querying pyre as top-level functions and modify test cases to call these functions.

Reviewed By: pradeep90

Differential Revision: D44669132

fbshipit-source-id: e240dd0e5faedec615203da2c43bb0ad6bc4952d