Update entrypoint-leaks to use top-level `find_issues`
Summary:
Context: We want to separate out callgraph utilities to allow other tools (for example to collect readonly adoption metrics) to use the same setup to walk through transitive callees given an entrypoint. The following set of diffs accomplish this separation.

This diff updates entrypoint-leaks command to use top-level `find_issues` function by removing the `find_issues` method from `CallGraph`. Now we collect a list of functions that we want to find issues for and pass that to `find_issues` top-level function.

Reviewed By: pradeep90

Differential Revision: D44716543

fbshipit-source-id: fe2c4d13eaf8575497567db6c9b70784c5cebc5f