Validate that actual == expected in AstEnvironmentTest.ml update tests
Summary:
Up until this diff, the AstEnvironment tests for update propagation were
not actually verifying that the triggered dependencies match, only a
one-way subset relation. That meant we would be unable to detect changes
of behavior where we triggered more dependencies than expected.

It turns out quite a number of the test cases were misleading - we were
explicitly stating in tests that wildcard dependencies *aren't* triggered
as invalidated sources, but this is actually wrong. They *are* triggered,
but because the test code didn't check for set containment symmetrically
the tests passed with incorrect expectations. All of the updated tests
have this flavor - some wildcard dependency triggers invalidation which
wasn't being propertly expressed in the expectations.

I have to fix this before I can push (with tests) an incremental bugfix.

Reviewed By: connernilsen

Differential Revision: D44797692

fbshipit-source-id: 63f4519d0bee1a1d2916a0871d2c579e132c7586