Emit errors when assigning to a global using setattr functions/methods
Summary:
**Background**: To ensure we're catching the more uncommon ways developers can mutate values, we want to make sure we're properly handling the builtins `setattr(...)` and `<anything>.__setattr__(...)`.

These two cases have special handling in `forward_expression`, since unlike other mutable expressions and assignment statements, the target global will be present in the arguments to the call rather than the target callee of the call expression.

Therefore, we check the first argument (the object being mutated) to see if there are reachable globals or errors returned from `forward_assignment`, and the third argument (the value the attribute of the argument is being set to) to see if there are any errors returned from `forward_assignment`. The second argument is the name of the attribute being set, which is not needed for this analysis.

The relevant tests here are moved into a new test function, `test_other_known_mutable_methods`, and descriptions are added for future maintainability.

Reviewed By: pradeep90

Differential Revision: D44722362

fbshipit-source-id: 9743f8f4bee92979ae5abc250dbf92835b4c4993