Make RawSources option-valued, fix new-source bug
Summary:
As discussed in the previous diff marking a test case as actually
proof of a bug, up until now AstEnvironment's RawSources was
result-valued, where it could only either store a parse failure
or an Ast - it had no way to store a marker indicating that the
source doesn't even exist.

As a result, we weren't able to distinguish between a source
for which we haven't yet attempted a lookup versus a source where
the lookup resulted in us declaring that no such qualifier exists.

This problem - the need for any lazy table to understand the distinction
between "I haven't looked this up yet" versus "I looked this up, it
doesn't exist" - is exactly why *every* "normal" environment table
is option-valued.

To accomplish the same thing in RawSources, we can make it store
an optional result rather than a direct result. This allows us to record
`None` whenever module tracker says this qualifier does not exist.

Once we've done this, the `RawSources.mem` lookup inside of the
definition of `reparse_module_paths` in `process_module_updates` will
successfully pick up that a newly-added source corresponding to some
previously-failed lookup needs to count as an invalidated module,
and be passed downstream.

NOTE: this will not fix the bug in T150184864 by itself, because
there's also a module tracker check happening inside of UGE that
needs to be removed.

Reviewed By: grievejia

Differential Revision: D44797691

fbshipit-source-id: 2f6a715013ff91d460207fb6ce3a30624992c817