introduce source/sink leaf ports
Summary:
Context: We're looking to introduce the same functionality of explicit leaf ports that include taint paths that Zoncolan has recently introduced (D44513889). The idea is to be able to identify flows by the leaf name + port more precisely than just saying "this flow originates at callable foo and goes to bar".

A previous method to accomplish this goal was by using subkinds, but they come with lots of additional analysis cost. fahndrich suggested to instead create port names that are more precise than the generated ones when users define taint, so let's do that!

SAPP will soon (tm) be able to parse these ports, it might already be able to do so.

Reviewed By: arthaud

Differential Revision: D44725465

fbshipit-source-id: 86fa32758c99da725eb8243547b2afcb852e08da