Remove use of `fold2_exn`.
Summary:
There's no behavior change. While the previous code was not harmful, the `when length a = b` made it seem like that was core to the business logic, when it was really a vague form of type checking (i.e., `x, y = 1, 2, 3`, which would be invalid anyway).

imo it's better to keep the match statement clean and catch `Unequal_lengths` within the block.

eafp

Reviewed By: grievejia

Differential Revision: D45018041

fbshipit-source-id: 43d30383ec3fedc060f3418b5ebdd8ab16852064