Make a more general assert_update function
Summary:
This commit rewrites tests in terms of a new update function
that is much more general, because it can handle updating multiple
sources and external sources.

There's a lot of code here, but the actual logical content of the tests is completely unchanged for now.

The new tests are more verbose, but I think they are also better
because they are far more explicit:
- before, the test expecations depended heavily on the name and location of the
  module ("test.py" in the local root) but didn't actually state that, you had
  to read the `assert_updates` function hundreds of lines above to realize
  this.
- also, we were previously using optional values for `original_source`
  and `new_source`. I can kind of see why, it made things concise, but it
  also made tests where a source was added or deleted *really* hard to follow,
  because the important thing ("there *are* no sources [before / after]") was
  not directly expressed! This new code forces you to explicitly pass an empty
  list of sources in such tests, which seems much clearer.

There are not yet new tests using external sources, I need to add those
in a stacked diff.

Reviewed By: grievejia

Differential Revision: D45015037

fbshipit-source-id: 69aa5ae841761938b6553f75f140a9dd36f19a2d