Create leak kind while accumumating error
Summary:
Context: For efficient use of developers time, we want to progressively make IG leak free, in tiers, based on risk levels of the leaks. For example, the highest level of risk include global leaks to mutable datastructures like dict/set/list. To enable this tiering we want to add different error codes based on the type of write. Currently we just have one error code (3100) that just says there is a global leak.

This diff adds the error kind information to `leaked_global` so that we know the category of error at creation time. `append_errors_for_globals` now takes as parameter the kind of error to construct, this will be useful in the upcoming diffs when we want to pass different parameters to different error category.

Reviewed By: connernilsen

Differential Revision: D45046498

fbshipit-source-id: 46c9d98c9acab77d6b6f21943b79683fdb3875ec