Codify globals writes in walrus operatror
Summary:
Context: For efficient use of developers time, we want to progressively make IG leak free, in tiers, based on risk levels of the leaks. For example, the highest level of risk include global leaks to mutable datastructures like dict/set/list. To enable this tiering we want to add different error codes based on the type of write. Currently we just have one error code (3100) that just says there is a global leak.

This diff adds `WriteToGlobalVariable` to assignments via the Walrus operator.

Reviewed By: connernilsen

Differential Revision: D45046580

fbshipit-source-id: 88b368f6c8cd80be9307641b7e3b647debced27c