Emit error when global is passed as a function parameter
Summary:
Context: Since we do not support interprocedural analysis, we would like to detect globals that are passed into functions as leaks as they can be modified in the calling function.
This diff updates the analysis for `Call` by erroring on all `reachable_globals`. Exisiting test cases are updated to reflect the change and new test cases are added to test the changes.

Reviewed By: connernilsen

Differential Revision: D44916131

fbshipit-source-id: 59857160cec4b84dbdb6d34aa5405feca2d50fdf