Codify return of globals
Summary:
Context: For efficient use of developers time, we want to progressively make IG leak free, in tiers, based on risk levels of the leaks. For example, the highest level of risk include global leaks to mutable datastructures like dict/set/list. To enable this tiering we want to add different error codes based on the type of write. Currently we just have one error code (3100) that just says there is a global leak.

This diff adds `ReturnOfGlobalVariable` error codes to globals that are returned from a function.

Reviewed By: connernilsen

Differential Revision: D45046573

fbshipit-source-id: 24f4af09d1a6bc68717f0395139ae55561910540