Codify direct writes to global variables
Summary:
Context: For efficient use of developers time, we want to progressively make IG leak free, in tiers, based on risk levels of the leaks. For example, the highest level of risk include global leaks to mutable datastructures like dict/set/list. To enable this tiering we want to add different error codes based on the type of write. Currently we just have one error code (3100) that just says there is a global leak.

In this diff I add a function to construct `WriteToGlobalVariable` error kind. Reachable globals from Assign target is modified to use this as the error kind. Further writes to globals are differentiated based on if its Primitive/Class/Mutatble Datastructure. Note that this diff only identifies direct writes to variables. The detection of class sometimes results in primitive. Will investigate and fix this in an upcoming diff.

Reviewed By: connernilsen

Differential Revision: D45046528

fbshipit-source-id: f0b24ddc4b419a25cce8a32f9005569f88403d82