Emit error when global is assigned to a local
Summary:
Context: We would like to detect globals that are assigned to local variables as they could be potential data leaks. Ideally we would also do taint analysis to only error on those locals with global references that are mutated. However, due to the low occurrence of such writes we can avoid this expensive analysis.

This diff updates the analysis for `Assign` by erroring on all `reachable_globals`. Exisiting test cases with todos are updated to reflect the change.

Reviewed By: connernilsen

Differential Revision: D44916148

fbshipit-source-id: 849d1dc51d7068e52f7c8cb948b72cb5c1d7b1ab