Add autocomplete query
Summary:
This is the first diff in a stack to implement basic autocomplete in the pyre code navigation server.

This diff implements the codenav query which is currently un-testable until we make a client API. That will come in future diffs.

The response payload currently only consists of a `label`. More attributes will be added such as kind, docstring (when available), and more shown in T151099563. I believe these attributes belong in the server response since the server has the best knowledge of what completion is what kind.

When "signature help" is implemented, an additional `SelectedCompletionInfo` API may be added here. But an argument can be made to only include that in our python client since we can deduce the necessary information from this `CompletionItem` API once adding the necessary attributes from T151099563.

Reviewed By: grievejia

Differential Revision: D45097393

fbshipit-source-id: f459541cc2ec79cec205770ef6ff7adb52b864e8