Add a test reproducing our new-external-module bug
Summary:
In D44797691 we fixed a bug in AstEnvironment where it didn't store enough
information to even know that we tried to lazy-load a module.

But the bug we wanted to fix is still there in our end-to-end flow. The reason
is that in addition to the AstEnvironment fix, we also need to change the way
the UGE lazy-loader works: today it bypasses AstEnvironment by directly asking
module tracker whether a file exist, which then means AstEnvironment's update
step doesn't know it should propagate updates on that module.

We'll fix the bug in the next diff.

Reviewed By: pradeep90

Differential Revision: D45089962

fbshipit-source-id: cdbdfb0d6898da8d5f90814244e514f33f2e27c3