Add new test of another incremental UGE bug
Summary:
At this point, we have fixed the bug where a combination
of AstEnvironment and UGE behaviors caused new *external*
modules to never be propagated in updates (as in, they were
completely missing from the `invalidated_modules` field of
the AstEnvironment update).

But we *still* have a bug: lazy loading is not currently
dependency-tracked. In most cases this is fine, because
module updates are usually 1:1 (as in, a changed module can only
change lookups of data in that module), and as a result the
multi-table code in UGE will trigger dependencies.

But in the special case of implicit modules, we get a bug because
lazy loading can cause lookups to short-circuit when no package
is found. When a module is added, we'll realize that there is
both a new implicit and explicit module, but because the dependency
table doesn't record lazy-loading deps, we fail to re-type-check
code that led to a short-circuited name lookup.

This diff adds a test illustrating where the problems pop up. I'll
fix the bug in the next diff.

Reviewed By: pradeep90

Differential Revision: D45091845

fbshipit-source-id: fe4fa1e6ce57f846fd36a5dade21ded6e9fd23c6