Add dependency tracking to lazy loading
Summary:
This commit halfway-fixes a bug where module metadata fetches of
implicit modules would lead to untracked dependencies, and then
adding new code in a previously-empty directory would produce
incremental bugs.

The partial fix here is that we add dependency tracking to the lazy
loading. This fixes lazy-loading bugs entirely, and indeed we see that
a previously incorrect unit test changed: one of our tests was asserting
that removing `test.py` wouldn't trigger a dependency on it which was
not correct.

But it doesn't fix the bug for module metadata loads; the reason is that
we are once again directly calling module tracker in the ReadOnly functions,
bypassing dependency tracking. I'll patch that in the next diff.

Reviewed By: pradeep90

Differential Revision: D45091843

fbshipit-source-id: 0cc8eb0b3eaefe9d93f7eb08acba4e8ada05ed13