Fix UGE incremental bug on new external modules
Summary:
By not going directly to ModuleTracker when checking for
existence of a module, but rather lettign AstEnvironment
always handle this, we fix an incremental bug.

The reason is that AstEnvironment update filters the full set of module updates
to only modules that are *either*:
- part of the project (this is needed so we know to typecheck new code)
- *or*, part of the RawSources map, i.e. some previous operation tried to load
  them.

Since D44797691, if we ask AstEnvironment for a module that doesn't exist we'll
record a failed lookup in RawSources, which tells AstEnvironment to tell UGE
about an update if the module appears later. That fix combined with this one
fixes the bug where new external modules were failing to propagate properly.

Reviewed By: pradeep90

Differential Revision: D45089961

fbshipit-source-id: 5e62d053bf0f953a74d7d09abf9928d5f3e0e595