Make sure all invalidated modules pass through lazy loading
Summary:
The final piece of the known incremental bugs in UGE is due to us having divergent
codepaths for loading modules that can bypass physically storing a marker in the
modules table (and thereby triggering dependencies) in lazy-load versus update.

The fix here is that update should pass *through* the lazy loader, so that lazy
loadign has sole responsibility for recording all data for both implicit and explicit
modules. This makes it easier to be confident that dependencies will be triggered
properly and indeed fixes the remaining bug (where removal of an implicit module might
not lead to re-checking code where the type check result depended on that implicit
module being found).

NOTE: as part of grievejia's investigation into how to make
Pyre and buck work better and learn from Hack, we've been thinking
about what a rewrite would look like. A big theme is that we
may want much coarser data - likely module-level in the "dependency-tracked"
part of our environment (I use scare quotes because the nature of dependency
tracking could change a lot of we re-architect - we may wind up with something
closer to an append-only setup that relies on merkle trees to uniquely identify
chunks of code). I think this whole chain of fixes (4 separate code-level bugs across
two environment layers in both the loading and updating phases resulting
in two categories of user-facing bug) should factor in: clearly our existing dependency
tracking is very complicated, I think it's near the limit of our ability to debug and fix. I'm
the person most familiar with this code and it took me several days to debug, a few more
to write regression tests.

Reviewed By: pradeep90

Differential Revision: D45091844

fbshipit-source-id: 162bf0bdec623486dbc3f01122c18b7461e8f7e5